Move mathesar types api tests; fix typo
