+++ mathesar/tests/api/test_database_api.py
@@ -2,8 +2,6 @@ import pytest
 from django.conf import settings
 from django.core.cache import cache
 
-from mathesar.api.display_options import DISPLAY_OPTIONS_BY_TYPE_IDENTIFIER
-from mathesar.api.dj_filters import FILTER_OPTIONS_BY_TYPE_IDENTIFIER
 from mathesar.reflection import reflect_db_objects
 from mathesar.models import Table, Schema, Database
 from db.tests.types import fixtures
@@ -88,7 +86,7 @@ def check_database(database, response_database):
     assert database.name == response_database['name']
     assert database.deleted == response_database['deleted']
     assert 'supported_types_url' in response_database
-    assert '/api/db/v0/databases/' in response_database['supported_types_url']
+    assert '/api/ui/v0/databases/' in response_database['supported_types_url']
     assert response_database['supported_types_url'].endswith('/types/')
 
 
@@ -203,58 +201,3 @@ def test_database_detail(client):
 
     assert response.status_code == 200
     check_database(expected_database, response_database)
-
-
-def test_type_list(client, test_db_name):
-    database = Database.objects.get(name=test_db_name)
-
-    response = client.get(f'/api/ui/v0/databases/{database.id}/types/')
-    response_data = response.json()
-    assert response.status_code == 200
-    assert len(response_data) == len(database.supported_types)
-    for supported_type in response_data:
-        assert all([key in supported_type for key in ['identifier', 'name', 'db_types', 'filters']])
-        found_filters = supported_type.get('filters')
-        expected_filters = FILTER_OPTIONS_BY_TYPE_IDENTIFIER.get(supported_type.get('identifier'))
-        assert found_filters == expected_filters
-        found_display_options = supported_type.get('display_options')
-        expected_display_options = DISPLAY_OPTIONS_BY_TYPE_IDENTIFIER.get(supported_type.get('identifier'))
-        assert found_display_options == expected_display_options
-
-
-def test_database_types_installed(client, test_db_name, engine_email_type):
-    expected_custom_types = [
-        {
-            "identifier": "email",
-            "name": "Email",
-            "db_types": [
-                "MATHESAR_TYPES.EMAIL"
-            ],
-            "filters": None,
-            'display_options': None
-        },
-        {
-            "identifier": "money",
-            "name": "Money",
-            "db_types": [
-                "MONEY",
-                "MATHESAR_TYPES.MONEY"
-            ],
-            "filters": None,
-            'display_options': None
-        },
-        {
-            "identifier": "uri",
-            "name": "URI",
-            "db_types": [
-                "MATHESAR_TYPES.URI"
-            ],
-            "filters": None,
-            'display_options': None
-        },
-    ]
-    reflect_db_objects()
-    default_database = Database.objects.get(name=test_db_name)
-
-    response = client.get(f'/api/ui/v0/databases/{default_database.id}/types/').json()
-    assert all([type_data in response for type_data in expected_custom_types])
